Add empty dashboard (#2)
* Fix env var name

* Set up dashboard page and spec

This page should just include a heading, and be located at the application root.

Back office users will have a range of actions they need to do, so we shouldn't throw them straight into the registration process.

* Use back office app template

This commit removes layout bits we don't need for internal use. The back office doesn't need content like the feedback link or the footer helpline.

* Include ApplicationHelper from engines

Without this, we are missing the method to determine the title, so the app throws an error.