Add users to back office (#3)
* Install Devise

Using the same version as waste-carriers-back-office for consistency.

* Configure Devise

This config should be the same as waste carriers.

* Run people migration from engine

This is the first time `rake db:migrate` was run since DEFRA/waste-exemptions-engine#39 was merged in, so the `people` model from the engine is added to the schema.

* Add user model and routes

This commit adds the database migration for a new User model, using Devise. It also adds the usual routes to the app.

* Add specs and factory for User

This commit adds some specs for the User model to test our password validation.

We were also misisng this support file for FactoryBot, which is needed for our factories.

* Run user DB migration

* Update Travis config

Migrate the DB before running the test suite.

* Enable Devise request helpers

These let us simulate users being signed in or signed out.

* Implement better session security

This allows us to invalidate user session cookies after logout, a security fix implemented in waste carriers in https://github.com/DEFRA/waste-carriers-back-office/pull/124. This prevents attackers from repeating requests made while logged in by reusing the cookie.

* Don't reload from cache on back button

This is a security measure designed to stop users from being able to access cached content with the back button after they log out.

* Users must be logged in to access back office

If a user who is not logged in attempts to access pages in the host app or the engine, they will be redirected to the Devise sign-in pages.

* Display user info on dashboard

Show the email address of the user who is logged in, plus a sign-out link.