Add registration search to dashboard
https://eaflood.atlassian.net/browse/RUBY-59

Logged-in users should be able to search for registrations. When the search term matches a registration, a basic overview of that registration should be displayed, along with actions the user can do.

This only covers completed registrations - searching for transient registrations (aka in-progress new registrations) will be added in a future PR.