Fix missing and inaccurate Devise text (#15)
We were missing some text used on the Devise login page. This commit adds the missing text so we don't get filler instead.

The page also referenced calling a helpline if you don't know your email - this was only intended for external users, so can be removed here.

In future, if text is missing from the translation files, the app will also raise an error so we can spot it more quickly.