Display registration details and actions
This commit adds the business name, reference number, location information and names of relevant people. It also lists all actions (including ones which may not be available) but these don't link to anything yet.

Following the design of the new waste-carriers-front-office dashboard as this is the most recent one which has had designer input.