Add registration search to dashboard (#9)
* Add registration search to dashboard

https://eaflood.atlassian.net/browse/RUBY-59

Logged-in users should be able to search for registrations. When the search term matches a registration, a basic overview of that registration should be displayed, along with actions the user can do.

This only covers completed registrations - searching for transient registrations (aka in-progress new registrations) will be added in a future PR.

* Add scopes for the search

The scopes check the following fields:

- site postcode
- registration reference number
- business or partner name
- applicant name
- applicant email
- contact name
- contact email

We do this by having one 'master' search scope which chains together the results of individual scopes on the registration itself and its relevant related objects.

Searches are case-insensitive. Queries on people's names concatenate the first and last name fields and then search the result, as otherwise a user searching 'Jane Smith' would not get matches for the separate fields - 'Jane' and 'Smith'.

These were initially added to the engine in https://github.com/DEFRA/waste-exemptions-engine/pull/47, but during review we agreed that actually scopes which will only ever be used in the back office should probably live there.

* Create SearchService

The dashboard will call an instance of this service to get the resulting matches.

If a term is provided, the Service uses scopes on the Registration to find matches. If no term is provided, return no results. (This also saves us from hitting the DB.)

* List search results on dashboard

The DashboardController gets the results from SearchService and displays them to the user.

A field on the dashboard provides the `term` param which is then used to make the search and display relevant results.

If there are no matches, we display a message saying that instead.

* Install and configure Kaminari

This is the gem we use for pagination. Config and styling should match waste carriers, with the exception of the Sass, which has been tidied up a bit to meet styling conventions.

We pass the page param from the DashboardsController to the SearchService so we can determine which results are shown on that page.

Pagination templates also appear on the dashboard when there are results to paginate.