Display regs on dashboard in a useful way (#16)
https://eaflood.atlassian.net/browse/RUBY-59

This PR covers the display and styling of registrations on the dashboard. Each registration should have key information clearly displayed, as well as what actions can be taken.

Each item on the dashboard shows the business name, reference number, location information and names of relevant people. It also lists all actions (including ones which may not be available) but these don't link to anything yet.

This follows the design of the new waste-carriers-front-office dashboard as this is the most recent one which has had designer input.