Add dashboard for user management (#19)
* Add dashboard for user management

https://eaflood.atlassian.net/browse/RUBY-69

This PR adds a dashboard listing all users in the system. Only system users should be able to access it.

This dashboard will serve as the launch point for managing users - changing their details or inviting new ones.

* Remove /bo from Devise paths

This was needed in waste carriers as we had to run two systems concurrently - however this won't be an issue for WEX so we can simplify the URL.

* Mount engine last in routes

Otherwise we run into clashing issues when we try to add more routes to the host app - it redirects to an error page instead.

* Add basic user dashboard

This follows the format from carriers - the dashboard is located at /users and lists an email address and role for each one.

* Only system users can view user dashboard

All others are redirected to a permissions error page.

* Display roles in a more user-friendly way

Use I18n to display a human-friendly role name, rather than the database value.

* Add pagination

* Add nav to top menu

Now that we have multiple pages in the app, we need some sort of top-level navigation. User management is only available to system users though, so only they should see the link.

* Display user status

Currently this is enabled by default, as RUBY-73 will add the ability to enable or disable a user.

* Add action links

Add links to change a user's role or disable them. (These features are yet to be implemented so links go nowhere.)

* Add link to invite users

Again, this feature will be implemented in a future story (RUBY-70) so currently link goes nowhere.