Add activate and deactivate pages
This commit adds a new controller for managing the activation and deactivation of users. Activating or deactivating a user goes through a confirmation page before the change is made. Attempting to access the page for an impossible action (eg deactivating a user who has already been deactivated) will redirect to the users list instead.

The actual activating and deactivating will be added in the next commit.