Fix password validation
Some tests threw up a validation error, which revealed that password validation hadn't been configured properly.

Devise handles password length validation for us with :validatable, so we don't need to add our own validation here.

When validating the format, the error message hadn't been set up properly, resulting in a translation error.

We also don't need to attempt to validate the format when no password is provided as Devise should handle this, so this commit adds a return clause to catch this. This matches what we did on old WEX and the advice from the Devise wiki: https://github.com/plataformatec/devise/wiki/How-To:-Set-up-simple-password-complexity-requirements