Make sure engine inherits Devise authentication
Additional testing revealed that the engine wasn't inheriting the authentication we'd set up using Devise – so a user who wasn't logged in could access pages in the engine, just not ones created in the host app.

This commit switches (temporarily) to a branch of the engine where this is fixed, and adds a test for it.