Block logged-in, deactivated users from accessing back office
If a user logs in but has been deactivated, they should not be able to view or use functionality in the back office host app or the engine. Instead, we redirect them to the page informing them that their account has been deactivated.

The exceptions are routes provided by Devise (used for signing in and out) and HighVoltage (used for providing static pages, including the one that says users are deactivated). We don't want to check abilities before the user has logged in, or stick deactivated users in a redirect loop when they are sent to the 'you're deactivated' page. So to prevent this, we check the controller before applying authorization and authentication checks.

This simulatenously fixes a bug which was calling `:authenticate_user!` too aggressively, blocking any user who wasn't logged in from trying to reset their password or unlock their account.