Configure Sendgrid and add rake task to test emails (#30)
https://eaflood.atlassian.net/browse/RUBY-150

The back office should use Sendgrid to send emails, although development defaults to Mailcatcher unless the correct env vars are supplied.

We also add a rake task to make sure this configuration is working correctly. Running `bundle exec rake email:test` should send out a simple test email to the address specified in the `EMAIL_TEST_ADDRESS` env var.

This commit also adds an initialiser for the engine. This existed in the front office but not the back office, and allows us to set some more config variables in a consistent way.