Activate and deactivate back office user accounts (#21)
https://eaflood.atlassian.net/browse/RUBY-73

System users should have the ability to activate or deactivate all back office user accounts. If a user logs into a deactivated account, they should see a message telling them the account is disabled. They should not be able to access any information or make any changes. Active accounts can access and use the system as normal.

This means we can deal with revoking access to the system without entirely deleting all record of the user (which will be important for version control, etc later on).

* Add 'active' column to users table

User accounts should be active by default.

* Add new methods to User

`active?` is a nicer way to check if a user is active, while `activate!` and `deactivate!` apply changes to the activity level.

* Display status activate/deactivate links on user dashboard

System users can now check a user's status and activate or deactivate them.

* Add activate and deactivate pages

This commit adds a new controller for managing the activation and deactivation of users. Activating or deactivating a user goes through a confirmation page before the change is made. Attempting to access the page for an impossible action (eg deactivating a user who has already been deactivated) will redirect to the users list instead.

* Block logged-in, deactivated users from accessing back office

The new :can_use_back_office ability relies on the user being active. We use this as a baseline check to see if users have permission to do things in the app or not.

If a user logs in but has been deactivated, they should not be able to view or use functionality in the back office host app or the engine. Instead, we redirect them to the page informing them that their account has been deactivated.

The exceptions are routes provided by Devise (used for signing in and out) and HighVoltage (used for providing static pages, including the one that says users are deactivated). We don't want to check abilities before the user has logged in, or stick deactivated users in a redirect loop when they are sent to the 'you're deactivated' page. So to prevent this, we check the controller before applying authorization and authentication checks.

This simultaneously fixes a bug which was calling `:authenticate_user!` too aggressively, blocking any user who wasn't logged in from trying to reset their password or unlock their account.

* Fix password validation

Some tests threw up a validation error, which revealed that password validation hadn't been configured properly.

Devise handles password length validation for us with :validatable, so we don't need to add our own validation here.

When validating the format, the error message hadn't been set up properly, resulting in a translation error.

We also don't need to attempt to validate the format when no password is provided as Devise should handle this, so this commit adds a return clause to catch this. This matches what we did on old WEX and the advice from the Devise wiki: https://github.com/plataformatec/devise/wiki/How-To:-Set-up-simple-password-complexity-requirements