Invite a new user to the back office
https://eaflood.atlassian.net/browse/RUBY-70

System users should be able to invite a new user to the WEX back office. When inviting a user, you must provide an email address and a role.

The invited user receives an email asking them to set a password so they can access the system. They must do this to complete the process and finish creating their account.