Update invitation acceptance page
When a user clicks the link in the invite email, they will be taken to this page and asked to enter a password.

Decided not to add tests as this doesn't modify the behaviour of the package at all, just the appearance.

Currently this throws an error as the newly-invited user has no role.