Invite a new user to the back office (#33)
https://eaflood.atlassian.net/browse/RUBY-70

System users should be able to invite a new user to the WEX back office. When inviting a user, you must provide an email address and a role. After sending an invite, system users are redirected to the user dashboard.

The invited user receives an email asking them to set a password so they can access the system. They must do this to complete the process and finish creating their account.

The devise_invitable gem was already installed before this PR – however, we needed to customise it to match the service's needs. This included improved content and applying our standard styling and structure.

We had to make some changes to our Devise config to make sure the role param is allowed and validated. To do this, we created a new UserInvitationsController to modify the behaviour of Devise (much like we do in SessionsController).

* Specs

Devise_invitable has its own tests. This commit adds some of our own specs for sending the invitations, as we had to build on top of the package to get the result we wanted (see UserInvitationsController).

However, we didn't add tests for accepting the invitation as we don't modify the behaviour of the package at all, just the appearance.

* Improved user email validation messages

This commit adds improved content for email validation error messages.

Validation on unique email addresses only kicks in when a user has accepted the invite. If a user is invited and does not accept, they can be invited a second time (to allow for resending missed emails).

* Adding 'invited' to status options on user dash

'Deactivated' takes precedence over other statuses. But if the user has a pending invitation and is active, we display 'invitation sent'.