Make use of airbrake configurable in initializer (#38)
Unfortunately the airbrake initializer errors if project_key is not set. The problem is that the initializer is fired in scenarios where we are not actually using the app, for example when running a rake task.

In production when we run rake tasks it's in an environment where environment variables have not been set. As such we need a way to disable using airbrake unless we actually need it.