Handle validation errors in form
Since the form comes with an option pre-selected, a user shouldn't really be submitting anything invalid unless they edit the HTML on the page or send their own POST request.

However, if such a thing does occur, we display the validation error on the form and don't persist the change.