Change a user's role (#52)
https://eaflood.atlassian.net/browse/RUBY-71

It should be possible to change a back office user's role. Only system users should be able to access this feature.

The role change form is prepopulated with the user's current role, but we also display it as text in case the form is updated and a reminder is then required.

This commit also refactors the 'invite a new user' form by moving the role selection into a partial. This means we can reuse it on the role change form and reduce duplication.