Add registration details page
https://eaflood.atlassian.net/browse/RUBY-60

This must work for both existing registrations, and registrations which are still being created (transient registrations).

The page should display all stored information about the registration, including its exemptions. It should also link to the actions which can be taken for individual exemptions or the entire registration.

Users must be logged in to access the page.