Link to start a new AD registration (#55)
https://eaflood.atlassian.net/browse/RUBY-61

Back office users should be able to start and complete a new registration.

* Add link to new registration

This link goes from the dashboard to the first form in the engine's registration journey.

* Implement sidebar pattern on dashboard

We use this sort of pattern on the waste carriers front office and back office dashboards. It does make things a bit more compact, and also provides access to the 'change password' link.

* Add feedback URL

The last page of the registration journey expects this value to be set, so without it, we get a translation error.