Add list of exemptions
The name of the relationship is different depending on whether we're dealing with a registration or a transient_registration, so a little extra logic is needed here.

Also move the display of the exemption data into its own partial to reduce duplication.

It might be better to modify the TransientRegistration model so it responds to `registration_exemptions`.