Block data agents from doing AD registrations (#61)
* Block data agents from doing AD registrations

https://eaflood.atlassian.net/browse/RUBY-61

Users with the role 'data_agent' should have read-only access. They shouldn't be able to create or modify new registrations. So if they try to access the user journey in the engine, they should instead be redirected to a permissions error page.

* Add permissions error page

We redirect users to this page if they try to do something which they don't have permission to do.

* Check permissions when users attempt to access engine

This check is skipped for static pages, so it should just affect forms in the registration journey.

* Hide AD registration link if users don't have permission