Search for in-progress new registrations (#62)
https://eaflood.atlassian.net/browse/RUBY-154

Logged-in users should be able to choose between searching existing registrations or ones which are still in progress (transient registrations). The results should be displayed in a way that is consistent with active registrations, with missing data dealt with gracefully.

Users can choose to search registrations or transient_registrations through radio buttons on the search form. Which button is preselected is determined by helper methods in the DashboardsHelper. The 'submitted' vs 'not yet submitted' wording for this is based on the old system.

The new `filter` parameter from the search form is then passed from the controller to the SearchService, which uses it to decide if it should call scopes on Registration or TransientRegistration.

This PR also refactors how the scopes used to get the search results work. Where possible, we've moved these into concerns to reduce duplication across the models. Methods which specifically call Address and Person have been left where they are, as we'll need to call TransientAddress and TransientPerson. It also standardises the use of `require` across the patched models, as it was inconsistent on the People model.

In the controller and view, we rename the variable storing the search results to`@results`, as this is now more accurate than `@registrations`. The HTML for the search results is also moved into a partial with more generic variable naming so it can work for both registrations and transient_registrations.

We also make some improvements to the information displayed with each result:

- Only display relevant links - some actions can only be performed for registrations or transient registrations, so we should only show them when relevant. (Although they're yet to be implemented anyway.)
- Display status of result - This should make it easier to distinguish between live and in-progress registrations. We use this flag pattern in the waste carriers back office already. Currently all registrations are marked as active - once we implement RUBY-64 and can deregister them, this will need to be updated.