Simplify logic to allow for new transient_registration aliases
It used to be that we had to call different methods to get the same type of data, depending on what object we were looking at - for example, `registration.addresses` and `transient_registration.transient_addresses`. However, now we have aliased these associations, so we can simplify the logic for displaying this data.