Deal with excess whitespace when searching (#71)
https://eaflood.atlassian.net/browse/RUBY-154

This fixes an issue flagged in QA of the transient reg search story. When a search term contains extra whitespace at the start or end of the term, this whitespace should be stripped.

Adding the test for stripping whitespace exposed that we weren't dealing well with paginating results when there were no results. So this commit also refactors the `search` method and breaks it down into smaller methods so we can respond more gracefully when no results are found.