Remove site address from results if not yet added (#74)
https://eaflood.atlassian.net/browse/RUBY-154

This fixes an issue flagged in QA of the transient reg search story. We decided it would be more consistent if this section just wasn't displayed when the data was missing (as with other sections in the results) rather than showing some filler text. However, the div still requires some contents, or it collapses. In this case a `<br>` seems to do the trick.