Add registration details page (#56)
https://eaflood.atlassian.net/browse/RUBY-60

This must work for both existing registrations, and registrations which are still being created (transient registrations).

The page should display all stored information about the registration, including its exemptions. It should also link to the actions which can be taken for individual exemptions or the entire registration.

Users must be logged in to access the page.

The routes follow the usual Rails pattern for a `show` action, so the page can be found at /registrations/[reference] and /transient-registrations/[reference]. We decided to go for a separate route for transient_registrations, as in the future a registration may coexist with a transient_registration - for example, if it's in the middle of being edited or renewed.

The two pages share most of the same content through a series of partial views.

This commit also refactors our tag SASS, as two different versions of it had been implemented around the same time (one in this PR and another in #62).