Add a root div to results in dashboard
This is a change purely to help the QA effort and writing our automated [acceptance tests](https://github.com/DEFRA/waste-exemptions-acceptance-tests). If the results can be wrapped in a root element it makes it easier to then identify and then work with the individual result elements, because we can build a section in our [page objects](https://github.com/natritmeyer/site_prism#sections).