Implement PaperTrail for auditing (#75)
https://eaflood.atlassian.net/browse/RUBY-203

PaperTrail is a gem used for keeping a record of when and how a record is changed. We use it to keep track of changes to back office users.

This PR installs and configures PaperTrail so we can migrate our existing audit history and continue to build on it. It adds and runs the migration for the new `versions` table in the database, and adds PaperTrail to the models we want to track (in this case, just Users).

The `whodunnit` for a new version should be set to the ID of the current_user who made the change.

The PaperTrail RSpec helper switches PaperTrail off by default when running tests (which should improve our speed) but also lets us use some extra matches when PaperTrail is enabled in individual specs.