Allow back office users to complete an in-progress registration
https://eaflood.atlassian.net/browse/RUBY-199

If an external user starts a registration but isn't able to complete it, and they call NCCC for help, NCCC staff should be able to pick it up and continue working on it.

The "Resume" links for this feature are already in place - this commit should make them do something useful.

Data agents should not be able to access this feature.