Remove extra version migration (#91)
This came out of our story to implement PaperTrail in this engine and the back office. Since we were tracking models in both projects, we ended up with two migrations to create a Versions table, which then caused a clash. So rather than have two migrations, it makes sense for the engine's migrations to take precedence, and for it to handle the creation of the versions table.

This PR removes the back office's version migration and updates to a version of the engine which should load its migrations before those of the host app (https://github.com/DEFRA/waste-exemptions-engine/pull/100).