Allow back office users to complete an in-progress registration (#85)
https://eaflood.atlassian.net/browse/RUBY-199

If an external user starts a registration but isn't able to complete it, and they call NCCC for help, NCCC staff should be able to pick it up and continue working on it. Data agents are not able to access this feature.

The "Resume" links for this feature were already in place - this commit makes them link to the current form in the registration journey.

This commit also adds a new ActionsLinkHelper to deal with the URLs for actions and the logic about whether or not to display those links.