Set default assistance_mode (#92)
https://eaflood.atlassian.net/browse/RUBY-200

A registration which is completed in the back-office should set the `assistance_mode` to `full` by default.

This commit runs the migration set up in https://github.com/DEFRA/waste-exemptions-engine/pull/98 and sets the correct value in the engine initializer.

It also displays assistance_mode on the registration details page. We include the text for `admin_intervention` and `partial` to deal with registrations migrated from the old schema, although it isn't possible to set those levels in this app yet.