Make whenever cron log output configurable
https://eaflood.atlassian.net/browse/RUBY-205

Should have spotted in PR #69 that the path set for the cron output is not available when the app is deployed to our AWS environment.

As what we have set in AWS also may not be suitable for local development this change makes the path configurable, though we default to what it would be in our AWS environments to simplify deployment.

---

WIP