Update schema to include deregistered_on
The engine recently had an additional migration added to include a `deregistered_on` date field against the exemptions.

However we are having issues in our environments with our db reset job because it is seeing an outstanding migration needs to be run.

So have run `db:migrate` to get the schema updated in order to resolve this issue.