Remove wkhtmltopdf-binary dependency
We haven't included this additional dependency when generating pdf's previously. It could be cool as it means we control our dependency on wkhtmltppdf rather than rely on the environment already having it installed.

However we couldn't use this particular one because it has no license (😡).

So to bring this in line with all our other projects, this change

- removes the wkhtmltopdf-binary gem
- adds the bash script used to start both xvfb and wkhtmltopdf (both are needed in our linux environments)
- updates the wickedpdf initializer to set wickedpdf's exe path as the path to the script