Update conf. letter presenter date_of_letter()
This change resolves the issue with the date not being formatted as the current letter.

Having reviewed the new code and the code from the previous app I can see they are exactly the same. Both use `I18n.l(Time.zone.today, format: :long)`.

However this results in different formats, probably as a result of differences in the environments where the PDF's are being generated.

As such this change brings formatting the date inline with how we have done it in other places, using the Rails `time_formats.rb` initializer. This should mean we are consistent across all environments now as well.