Fix description of exemption
Having compared the code across the old and new service you can see why this has got out of whack.

The old code refers to what looks like an attribute called `exemption_description`. What this actually does is call a custom `to_s` method on the `exemption` model which returns the code and the summary. So calling `.description` is a logical choice but it actually returns the full description for the exemption.

If we have a use for a similarly formatted string in other parts of the service, the logic I've implemented in the view could also be moved to the `Exemption` model in the engine.

Till then though I believe this is perfectly sufficient for our needs now, and updates the letter to match the current service.