Move exemption description logic into presenter
This is actually because of the next fix that is coming, which is to display the expiry date of the exemption rather than just its status in order to match the current letter.

Upon review it turns out the existing code has a bunch of helpers built into its `EnrollmentExemption` which return a string based on the status of the exemption. For example if active "Expires on foo", if expired "Expired on foo", and if revoked "Revoked".

This means a bunch more logic and as we haven't got a reuse case, there is no call to build it into the `RegistrationExemption` yet. However we don't want it in the view so next best is the presenter.

And if we are adding that logic into the presenter, why not add the description logic? So as a precursor to that change, this change moves the logic for generating a exemption description into the presenter, and creates a spec so we can start testing these new methods.