Use reg reference in edit URLs to find or create EditedRegistration
Our forms generally rely on the token of the TransientRegistration. If there's no token, we make a new TransientRegistration. However, when we initialize an EditedRegistration, we need an existing Registration to base it on.

This change means that including an existing Registration's reference in the edit form URL can be used to find or initialize an EditedRegistration.