Add form test for 100% coverage
Initially I didn't include any tests for the edit forms as they aren't being used to save anything, so are covered adequately by the request tests. We've done the same thing for other similar form objects in the engine.

However, the code coverage report revealed that the EditCompleteForm's submit method wasn't being called, and so wasn't being tested. So this commit adds a test for it (and gets CodeClimate off my back).