Fix status description for exemptions
So this is the actual fix for the exemption status field. Prior to this we were just displaying the state, however if you check the existing confirmation letter it actually displays a specific message depending on the state.

So for example if the registration exemption is active the current letter displays "Expires on 15 April 2022". If ceased it displays "Ceased on 10 April 2019" (same for revoked, just the status changes).

Because of this extra logic I've placed it in the presenter, and added more tests to the spec.