Set up edit form with data and links
This still needs some fine-tuning but all the info is there.

I had to deviate from Liam's design in the prototype - the appearance is similar but I'm using tables instead of definition lists. The prototype uses the new design system and unfortunately it wasn't clear how I could backport the summary list SCSS without a lot of work. So it's an elements-friendly table instead.