Use NewRegistration instead of TransientRegistration when appropriate (#141)
https://eaflood.atlassian.net/browse/RUBY-62

The engine now allows for multiple types of TransientRegistrations, starting with a NewRegistration.

The back office has generally been assuming that TransientRegistrations are only for the new registration journey. Now that we have the NewRegistration class specifically for that, we need to update the back office to use it.

This includes:

- Changing the TransientRegistration page to NewRegistration page
- Search NewRegistrations instead of TransientRegistrations
- Fix action links
- Update abilities
- Update pending tag on dashboard

It also runs a database migration for the type.