Fix EditHelper inheritance and overriding (#151)
https://eaflood.atlassian.net/browse/RUBY-62

Previously we were replacing the engine's EditHelper with our own. However, this meant that any new methods added to the original EditHelper were not accessible. Instead, we should make a new EditHelper for the back office and just include the original EditHelper. This lets us override certain methods but keep the others.