Fix bug on show of New registration objects
Fixes https://eaflood.atlassian.net/browse/RUBY-329

The issues was related to a missing scope for transient registration exemptions.
It was only appearing for in-progress registration with already related transient exemptions.
In order to reproduce the bug on the test suite and prevent regression, exemptions have been added to the base new_registration object.
The missing scope have been externalised to a concern and included in transient registration exemptions models from the engine.