Simplify the view and presenter - Registration details
I know this is as a result of copying and pasting the code from the existing Waste Permits project, but currently IMHO we have an unnecessarily complex presenter to support a bit of reuse in the view.

So this is the first of a few commits that is working to drop the `_letter_section.html.erb` partial and the supporting, overly complex methods in the presenter.