Adopt base_presenter delegation pattern for letter
In other projects we have use a `base_presenter` essentially as a way of merging the attributes of the source model and the presenter.

Prior to this change `ConfirmationLetterPresenter` held the registration as an attribute, and explicitly delegated calls to `reference` to it.

As we are now intending to make calls to a number of other registration attributes as part of our simplification work we have 3 options

- expose `@registration` as an attribute on the presenter and then call `confirmation_letter_data.registration.some_attribute` in the views
- add a `delegate :some_attribute, to: :registration` for every attribute we wish to expose
- adopt the `base_presenter` pattern which uses a `SimpleDelegator`

Obviously with this change I have gone with No. 3, as I believe

- it makes the letter presenter consistent with other projects
- it is much more flexible and less brittle than option 2
- it feels cleaner than option 1

Essentially it gets the presenter closer to what I believe the purpose of a presenter is; a resource that holds both the model data and the logic for displaying that model data needed by the view consuming it.